QDB-12511- Investigate cluster connection closing / not closing (#47)
This adds checks in a lot of places that explicitly checks whether the connection to the cluster is still open. Additionally, we put the same check in some strategic places (e.g. where we convert a qdb::handle to qdb_handle_t).

Also adds an explicit exception type for quasardb.InvalidHandleError.