QDB-12569 - Measure low-level C API timings (#48)
This implements a "quasardb.metrics" module which is mostly intended for internal use, to keep track of where time is spent. It may be useful in future scenarios to debug performance issues such as "is time spent in the C API or in Python doing conversions".

From the Python side, the metrics module does expose a few functions directly (such as quasardb.metrics.totals() and quasardb.metrics.clear()), but these aren't really intended to be used directly. They may become useful in the future if we want to hook some Cloudwatch exporter or something into this, but this is out of scope rigth now.

The intended usage is through the metrics.Measure class as follows:

```
with metrics.Measure() as measure:
   # do_stuff
   timings = measure.get()
   total_ns_spent_query = timings["qdb_query"]
   total_ns_spent_push = timings["qdb_ts_batch_push"]
```

Timings is just a regular python dict, with a metric id as key, and total nanoseconds as value.