QDB-15011 - Flaky python tests due to setuptools (#77)
Setuptools' built-in integration of testing has been deprecated; this bug started surfacing more and more with recent versions of Python.

Since Python 3.12, distutils officially has officially been deprecated, and setuptools still ships with a shadow copy of it, but they're slowly removing this.

This patch moves some things around. Specifically:
- rather than running python3 setup.py test, we now invoke python3 -m pytest directly;
- we are super careful to execute all commands inside a virtualenv, which did not work properly before on windows;
to avoid conflicts between the directory quasardb and the installed quasardb module, we execute pytest in the tests/ subdirectory;
- we now use the build python module to handle the building of the python api for us.

Still left to-do:
- switch documentation backend, it's deprecated and doesn't work properly anymore;